chore: add schema only if installed via `node_modules`
