+++ crates/rome_cli/tests/commands/mod.rs
@@ -1,5 +1,6 @@
 mod check;
 mod ci;
 mod format;
+mod init;
 mod rage;
 mod version;
