+++ crates/rome_cli/tests/configs.rs
@@ -9,6 +9,15 @@ pub const CONFIG_FORMAT: &str = r#"{
 "#;
 
 pub const CONFIG_INIT_DEFAULT: &str = r#"{
+  "linter": {
+    "enabled": true,
+    "rules": {
+      "recommended": true
+    }
+  }
+}"#;
+
+pub const CONFIG_INIT_DEFAULT_WHEN_INSTALLED: &str = r#"{
   "$schema": "./node_modules/rome/configuration_schema.json",
   "linter": {
     "enabled": true,
