+++ crates/rome_cli/tests/main.rs
@@ -217,52 +217,6 @@ mod main {
     }
 }
 
-mod init {
-    use super::*;
-    use crate::configs::CONFIG_INIT_DEFAULT;
-    use crate::snap_test::SnapshotPayload;
-    use pico_args::Arguments;
-    use rome_console::BufferConsole;
-    use rome_fs::{FileSystemExt, MemoryFileSystem};
-    use rome_service::DynRef;
-    use std::ffi::OsString;
-    use std::path::Path;
-
-    #[test]
-    fn creates_config_file() {
-        let mut fs = MemoryFileSystem::default();
-        let mut console = BufferConsole::default();
-
-        let result = run_cli(
-            DynRef::Borrowed(&mut fs),
-            DynRef::Borrowed(&mut console),
-            Arguments::from_vec(vec![OsString::from("init")]),
-        );
-        assert!(result.is_ok(), "run_cli returned {result:?}");
-
-        let file_path = Path::new("rome.json");
-
-        let mut file = fs
-            .open(file_path)
-            .expect("configuration file was not written on disk");
-
-        let mut content = String::new();
-        file.read_to_string(&mut content)
-            .expect("failed to read file from memory FS");
-        assert_eq!(content, CONFIG_INIT_DEFAULT);
-
-        drop(file);
-
-        assert_cli_snapshot(SnapshotPayload::new(
-            module_path!(),
-            "creates_config_file",
-            fs,
-            console,
-            result,
-        ));
-    }
-}
-
 mod configuration {
     use super::*;
     use crate::configs::{
