+++ editors/vscode/configuration_schema.json
@@ -857,6 +857,7 @@
           ]
         },
         "useConst": {
+          "description": "Require const declarations for variables that are never reassigned after declared.",
           "anyOf": [
             {
               "$ref": "#/definitions/RuleConfiguration"
