+++ npm/rome/configuration_schema.json
@@ -563,6 +563,17 @@
       "description": "The configuration of the filesystem",
       "type": "object",
       "properties": {
+        "ignore": {
+          "description": "A list of Unix shell style patterns. Rome tools will ignore files/folders that will match these patterns.",
+          "type": [
+            "array",
+            "null"
+          ],
+          "items": {
+            "type": "string"
+          },
+          "uniqueItems": true
+        },
         "maxSize": {
           "description": "The maximum allowed size for source code files in bytes. Files above this limit will be ignored for performance reason. Defaults to 1 MiB",
           "default": null,
@@ -797,6 +808,17 @@
         },
         "useCamelCase": {
           "description": "Enforce camel case naming convention.",
+          "anyOf": [
+            {
+              "$ref": "#/definitions/RuleConfiguration"
+            },
+            {
+              "type": "null"
+            }
+          ]
+        },
+        "useConst": {
+          "description": "Require const declarations for variables that are never reassigned after declared.",
           "anyOf": [
             {
               "$ref": "#/definitions/RuleConfiguration"
